+++ src/ElectroMagn/LaserEnvelope.h
@@ -20,7 +20,8 @@ class LaserEnvelope
 public:
     LaserEnvelope( Params &params, Patch *patch ); // Main constructor
     LaserEnvelope( LaserEnvelope *envelope, Patch *patch, Params &params, unsigned int n_moved ); // Cloning constructor
-    virtual void initEnvelope( Patch *patch, ElectroMagn *EMfields ) = 0;
+    virtual void initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields ) = 0;
+    virtual void injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual ) = 0;
     virtual ~LaserEnvelope();
     virtual void updateEnvelope( Patch *patch ) = 0;
     virtual void updateEnvelopeReducedDispersion( Patch *patch ) = 0;
@@ -41,6 +42,7 @@ public:
     // This coefficient is used for the ponderomotive potential Phi = ellipticity_factor*|A|^2/2.
 
     std:: string envelope_solver  = "explicit"; // default value
+    std:: string envelope_initialization = "inside_the_window"; // default value
     
     Field *A_;         // envelope value at timestep n
     Field *A0_;        // envelope value at timestep n-1
@@ -90,7 +92,8 @@ class LaserEnvelope1D : public LaserEnvelope
 public:
     LaserEnvelope1D( Params &params, Patch *patchs );
     LaserEnvelope1D( LaserEnvelope *envelope, Patch *patch, Params &params, unsigned int n_moved );
-    void initEnvelope( Patch *patch, ElectroMagn *EMfields ) override final;
+    void initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields ) override final;
+    void injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual ) override final;
     ~LaserEnvelope1D();
     void updateEnvelope( Patch *patch ) override final;
     void updateEnvelopeReducedDispersion( Patch *patch ) override final;
@@ -106,7 +109,8 @@ class LaserEnvelope2D : public LaserEnvelope
 public:
     LaserEnvelope2D( Params &params, Patch *patch );
     LaserEnvelope2D( LaserEnvelope *envelope, Patch *patch, Params &params, unsigned int n_moved );
-    void initEnvelope( Patch *patch, ElectroMagn *EMfields ) override final;
+    void initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields ) override final;
+    void injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual ) override final;
     ~LaserEnvelope2D();
     void updateEnvelope( Patch *patch ) override final;
     void updateEnvelopeReducedDispersion( Patch *patch ) override final;
@@ -122,7 +126,8 @@ class LaserEnvelope3D : public LaserEnvelope
 public:
     LaserEnvelope3D( Params &params, Patch *patch );
     LaserEnvelope3D( LaserEnvelope *envelope, Patch *patch, Params &params, unsigned int n_moved );
-    void initEnvelope( Patch *patch, ElectroMagn *EMfields ) override final;
+    void initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields ) override final;
+    void injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual ) override final;
     ~LaserEnvelope3D();
     void updateEnvelope( Patch *patch ) override final;
     void updateEnvelopeReducedDispersion( Patch *patch ) override final;
@@ -138,7 +143,8 @@ class LaserEnvelopeAM : public LaserEnvelope
 public:
     LaserEnvelopeAM( Params &params, Patch *patch );
     LaserEnvelopeAM( LaserEnvelope *envelope, Patch *patch, Params &params, unsigned int n_moved );
-    void initEnvelope( Patch *patch, ElectroMagn *EMfields ) override final;
+    void initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields ) override final;
+    void injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual ) override final;
     ~LaserEnvelopeAM();
     void updateEnvelope( Patch *patch ) override final;
     void updateEnvelopeReducedDispersion( Patch *patch ) override final;