+++ src/ElectroMagn/LaserEnvelope1D.cpp
@@ -68,7 +68,7 @@ LaserEnvelope1D::LaserEnvelope1D( LaserEnvelope *envelope, Patch *patch, Params 
 }
 
 
-void LaserEnvelope1D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
+void LaserEnvelope1D::initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields )
 {
     cField1D *A1D          = static_cast<cField1D *>( A_ );
     cField1D *A01D         = static_cast<cField1D *>( A0_ );
@@ -130,6 +130,27 @@ void LaserEnvelope1D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
     
 }
 
+void LaserEnvelope1D::injectEnvelopeFromXmin( Params &params, bool inject_envelope_from_this_patch, double time_dual )
+{
+    cField1D *A1D                 = static_cast<cField1D *>( A_ );
+    cField1D *A01D                = static_cast<cField1D *>( A0_ );
+    
+    double t                      = time_dual;          // x-ct     , t=0
+    double t_previous_timestep    = time_dual-timestep; // x-c(t-dt), t=0
+  
+    vector<double> position( 1, 0 );
+    position[0]                   = 0.;
+    
+    // oversize
+    int oversize_                 = params.oversize[0];
+    
+    // Impose the envelope value for x=0 at time t and t-dt 
+    if ( inject_envelope_from_this_patch ){    
+        ( *A1D  )( oversize_-1 ) += profile_->complexValueAt( position, t );
+        ( *A01D )( oversize_-1 ) += profile_->complexValueAt( position, t_previous_timestep );
+    }
+    
+}
 
 LaserEnvelope1D::~LaserEnvelope1D()
 {