+++ doc/Sphinx/Overview/releases.rst
@@ -35,6 +35,10 @@ Changes made in the repository (not released)
   * New argument ``timestep_indices`` as an alternative to ``timesteps``.
   * Changed coordinate reference for 2D probe in 3D or AM geometry
     (zero is the box origin projected orthogonally on the probe plane).
+    
+* **LaserEnvelope model**:
+
+  * Now the laser pulse can be initialized either inside the window (default mode) or from the ``xmin`` border setting the ``envelope_initialization`` variable.
 
 * Documentation: